backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
firebase-admin==6.3.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pinecone-client==5.0.1
boto3==1.34.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-json-logger==2.0.7

# LLM Evaluation Pipeline
wandb>=0.16.0
